backend/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
SQLAlchemy
PyMySQL
